+++ lib/recurly/webhook/scheduled_payment_notification.rb
@@ -1,0 +1,6 @@
+module Recurly
+  module Webhook
+    class ScheduledPaymentNotification < TransactionNotification
+    end
+  end
+end