+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## Unreleased
 
 * Add `cc_emails` attribute to `Account` [PR](https://github.com/recurly/recurly-client-ruby/pull/216)
+* Add webhooks parsers [PR](https://github.com/recurly/recurly-client-ruby/pull/217)
 
 <a name="v2.4.7"></a>
 ## v2.4.7 (2015-10-02)